Update /api/pdf route to use the new pdf rendering
