+++ server/services/__mocks__/pdfGenerationService.ts
@@ -1,0 +1,12 @@
+const mockGeneratePdf = jest.fn().mockResolvedValue(undefined)
+
+class PdfGenerationService {
+  constructor() {
+    // Empty constructor - preSentenceToDeliusService parameter is ignored in mock
+  }
+
+  generatePdf = mockGeneratePdf
+}
+
+export default PdfGenerationService
+export { mockGeneratePdf }
